requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.135.1
filelock==3.25.2
google-auth
google-auth-oauthlib
motor
requests
fsspec==2026.2.0
groq==1.1.1
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
openai==2.28.0
packaging==26.0
pandas==3.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pyttsx3==2.99
PyYAML==6.0.3
regex==2026.2.28
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.15.1
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.41.0
PyPDF2==3.0.1
python-docx==1.1.2
python-jobspy
pymongo
